feat: Add share password feature
This commit will add a share password feature, which will create the option
that a logged in user can share a password with another registered user.
This is done via a link on each password view.
The current_user is excluded from the dropdown list, and once a password
is shared with another user, that user is also excluded.